pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["custom_components/openinverter"]


[project]
name = "hass-open-inverter-gateway"
version = "0.1.1"
description = "Home Assistant integration for Open Inverter Gateway"
readme = "README.md"
requires-python = ">=3.12"
dependencies = ["aiohttp>=3.9.0"]

[project.scripts]
test = "pytest:console_main"

[dependency-groups]
dev = [
    "commitizen>=4.13.0",
    "homeassistant-stubs",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
    "ty>=0.0.14",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]


[tool.ty]
# Ty configuration (Astral's type checker)
# Assuming standard defaults for now, can be tuned later.
# Ty reads many settings from pyproject.toml automatically.

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true
